Add shortcut keys for Gamma
This commit will add shortcut keys to the gamma controls, rename some functions to be more organized, and a legend for the keys on the menu. These keys may work on non-debug builds, so beware of releasing the code without any modifications if you do not want undocumented features. The default shortcut keys are command option minus and command option equals, but the legend will show command option plus instead of equals for visual consistency, despite command option equals needing no shift key to press. This code was mostly copy and paste changes so should be easily modifiable for both the legend and the actual shortcut.